FDE-v21-b2

+++ system/engine/gears/sleeper.sh
@@ -2,41 +2,45 @@
 ### FeraDroid Engine v0.21 | By FeraVolt. 2016 ###
 B=/system/engine/bin/busybox
 LOG=/sdcard/Android/FDE.txt
-if [ -e /sys/power/wait_for_fb_sleep ]; then
- A=$(cat /sys/power/wait_for_fb_sleep)
- B=sleeping
- C=awake
-elif [ -e /sys/class/graphics/fb0/show_blank_event ]; then
- A=$(cat /sys/class/graphics/fb0/show_blank_event)
- B=panel_power_on = 0
- C=panel_power_on = 1
-fi;
-$B cp /sdcard/Android/sleeper_whitelist.txt /system/engine/assets/sleeper_whitelist.txt
 sync;
-$B mount -o remount,ro /system
+if [ -e /sdcard/Android/sleeper_whitelist.txt ]; then
+ $B mount -o remount,rw /system
+ $B rm -f /system/engine/assets/sleeper_whitelist.txt
+ $B cp /sdcard/Android/sleeper_whitelist.txt /system/engine/assets/sleeper_whitelist.txt
+ $B mount -o remount,ro /system
+fi;
 W=$($B cat /system/engine/assets/sleeper_whitelist.txt | grep -v -e '#' | tail -n1)
 ON=$($B cat /system/engine/assets/sleeper_whitelist.txt | grep -e 'sleeper=1')
 if [[ "sleeper=1" = "$ON" ]]; then
- $B echo "Sleeper daemon is active." > $LOG
+ $B echo "Sleeper daemon is active." >> $LOG
  while true; do
-  if [[ "$A" = "$B" ]]; then
+  if [[ "false" = "$(dumpsys power | $B grep -E "mScreenOn" | $B grep -o "false")" ]]; then
    sync;
-   sleep 9
+   for i in $($B pgrep -l '' | $B grep -E "org.|app.|com.|net.|eu." | $B grep -v -i -E "$W" | $B awk '{print $1}'); do
+     kill -9 "$i"
+   done;
    if [ -e /proc/sys/vm/drop_caches ]; then
+    sync;
     $B sleep 1
     $B echo 3 > /proc/sys/vm/drop_caches
     $B sleep 1
     sync;
    fi;
-   for i in $($B pgrep -l '' | $B grep -E 'org.|app.|com.|android.|net.|eu.' | grep -v -i -e $W  | $B awk '{print $1}'); do
-     kill -9 "$i"
+   if [ "$($B cat /sys/class/power_supply/battery/capacity)" -lt 30 ]; then
+    svc wifi disable
+    svc nfc disable
+    svc data disable
+    service call bluetooth_manager 8
+    setprop ro.com.google.networklocation 0
+    am broadcast -a android.intent.action.BATTERY_LOW
+   fi;
+   until [[ "true" = "$(dumpsys power | $B grep "mScreenOn" | $B grep -o "true")" ]]; do
+    $B sleep 600
    done;
-   $B sleep 27
-   until [[ "$A" = "$C" ]]; do
-    $B sleep 9
-   done;
+  else
+   $B sleep 18
   fi;
  done;
 else
- $B echo "Sleeper daemon is not active." > $LOG
+ $B echo "Sleeper daemon is not active." >> $LOG
 fi;
